Move punyCodeAndNormaliseUrl from Facade to UrlUtils
